requirements.txt
langchain
langchain-google-genai
langgraph
pydantic
pydantic-settings
python-dotenv
pytest
